+++ tests/Feature/InstallCommandTest.php
@@ -21,6 +21,8 @@ class InstallCommandTest extends TestCase
     protected function tearDown(): void
     {
         @unlink($this->envPath);
+        @unlink(public_path('vendor/warden/warden.css'));
+        @rmdir(public_path('vendor/warden'));
         parent::tearDown();
     }
 
@@ -50,6 +52,16 @@ class InstallCommandTest extends TestCase
             ->assertSuccessful();
 
         $this->assertStringContainsString('WARDEN_MODE=parent', (string) file_get_contents($this->envPath));
+    }
+
+    public function test_parent_install_publishes_the_dashboard_stylesheet(): void
+    {
+        @unlink(public_path('vendor/warden/warden.css'));
+
+        $this->artisan('warden:install', ['--parent' => true, '--no-migrate' => true])
+            ->assertSuccessful();
+
+        $this->assertFileExists(public_path('vendor/warden/warden.css'));
     }
 
     public function test_passing_both_flags_fails(): void